+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * fixed a bug in woebin using chimerge method
 * gains_table supports setting of break width
+* fixed a bug in germancredit data set
+*** scorecard2 supports adjusting sampling and return predict probability
 
 # scorecard 0.2.7